Improve handling of German ck -> kk hyphenation
Pre the German orthography reform of 1996, "ck" became "kk" during
hyphenation.  Let's change the "kk" back to "ck" so we can compare
the hyphenation pattern with the original word.

We have to do this because, for example, in the word "Muckefuck"
only the "ck" in "Mucke" becomes kk where "fuck" is not hyphened
at all and therefore stays "ck".